Serving/tests/requirements-test.txt
locust
pillow
onnxscript
